[UI Toolkit] Drop abstract theme classes in favor of interfaces
Interfaces have a clearer separation of concerns and are more flexible
than abstract classes for our themes. Migrating them doesn't really
change the game plan so we might as well do it!
