--- Assets/Shopify/UIToolkit/Themes/Interfaces.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: faba619e64e4448ae83ec593c0fd8d16
-folderAsset: yes
-timeCreated: 1511470193
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
