--- Assets/Shopify/UIToolkit/Themes/Interfaces/ISingleProductTheme.cs
@@ -1,7 +1,0 @@
-﻿namespace Shopify.UIToolkit {
-    using Shopify.Unity;
-
-    public interface ISingleProductTheme : IThemeBase {
-        void OnShouldShowProduct(Product product, ProductVariant[] variants);
-    }
-}
