--- Assets/Shopify/UIToolkit/Themes/Interfaces/ISingleProductTheme.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5c2f979ef242b40719fdb9ace2afc7d7
-timeCreated: 1511470380
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
