--- Assets/Shopify/UIToolkit/Themes/Interfaces/IThemeBase.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4b83f7b0b3b1b4442a616ed1b46215d0
-timeCreated: 1511470201
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
